# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CodeClash - Real-time competitive coding battle platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /battles/{battleId}:
    get:
      operationId: getBattle
      tags: [battles]
      summary: Get battle details
      parameters:
        - name: battleId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Battle"
  /battles/{battleId}/leaderboard:
    get:
      operationId: getBattleLeaderboard
      tags: [battles]
      summary: Get battle leaderboard
      parameters:
        - name: battleId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BattleParticipantStatus"
  /battles/{battleId}/submit:
    post:
      operationId: recordSubmission
      tags: [battles]
      summary: Record a Codeforces submission attempt
      parameters:
        - name: battleId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmissionRecord"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BattleParticipantStatus"

  # ── Problems ──
  /problems:
    get:
      operationId: listProblems
      tags: [problems]
      summary: List cached Codeforces problems
      parameters:
        - name: rating
          in: query
          schema:
            type: integer
        - name: topic
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Problem"
  /problems/random:
    get:
      operationId: getRandomProblem
      tags: [problems]
      summary: Get a random problem matching criteria
      parameters:
        - name: rating
          in: query
          schema:
            type: integer
        - name: topic
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
...
  /friends/request:
    post:
      operationId: sendFriendRequest
      tags: [friends]
      summary: Send a friend request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FriendRequestInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FriendEntry"
  /friends/{friendshipId}/respond:
    post:
      operationId: respondFriendRequest
      tags: [friends]
      summary: Accept or reject a friend request
      parameters:
        - name: friendshipId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FriendResponse"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FriendEntry"

  # ── Achievements ──
...
components:
  schemas:
...
    Battle:
      type: object
      required: [id, roomId, roomCode, battleType, status, problem, participants, startedAt]
      properties:
        id:
          type: integer
        roomId:
          type: integer
        roomCode:
          type: string
        battleType:
          type: string
        status:
          type: string
          enum: [active, finished]
        problem:
          $ref: "#/components/schemas/Problem"
        participants:
          type: array
          items:
            $ref: "#/components/schemas/BattleParticipantStatus"
        startedAt:
          type: string
        endedAt:
          type: ["string", "null"]
        winnerId:
          type: ["integer", "null"]
        winnerUsername:
          type: ["string", "null"]
        durationSeconds:
          type: ["integer", "null"]
    BattleParticipantStatus:
      type: object
      required: [userId, username, cfHandle, rating, rank, solved, attempts, penalty, rank_position]
      properties:
        userId:
          type: integer
        username:
          type: string
        cfHandle:
          type: string
        rating:
          type: integer
        rank:
          type: string
        solved:
          type: boolean
        solveTimeSeconds:
          type: ["integer", "null"]
        attempts:
          type: integer
        penalty:
          type: integer
        rank_position:
          type: integer
        isEliminated:
          type: boolean
        lastVerdict:
          type: ["string", "null"]
    SubmissionRecord:
      type: object
      required: [cfSubmissionId, verdict]
      properties:
        cfSubmissionId:
          type: integer
        verdict:
          type: string
          enum: [OK, WRONG_ANSWER, RUNTIME_ERROR, TIME_LIMIT_EXCEEDED, COMPILATION_ERROR, MEMORY_LIMIT_EXCEEDED, PARTIAL]

    # ── Problems ──
    Problem:
      type: object
      required: [id, contestId, problemIndex, title, rating, tags, cfUrl]
      properties:
        id:
          type: integer
        contestId:
          type: integer
        problemIndex:
          type: string
        title:
          type: string
        rating:
          type: integer
        tags:
          type: array
          items:
            type: string
        cfUrl:
          type: string
        solvedCount:
          type: ["integer", "null"]
...
    FriendEntry:
      type: object
      required: [id, userId, friendId, friendUsername, friendCfHandle, friendRating, friendRank, status]
      properties:
        id:
          type: integer
        userId:
          type: integer
        friendId:
          type: integer
        friendUsername:
          type: string
        friendCfHandle:
          type: string
        friendRating:
          type: integer
        friendRank:
          type: string
        status:
          type: string
          enum: [pending, accepted, rejected]
        isOnline:
          type: boolean
    FriendRequestInput:
      type: object
      required: [targetUsername]
      properties:
        targetUsername:
          type: string
    FriendResponse:
      type: object
      required: [accepted]
      properties:
        accepted:
          type: boolean

    # ── Achievements ──
...