# lib/api-zod/src/generated/types/friendEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CodeClash - Real-time competitive coding battle platform
 * OpenAPI spec version: 0.1.0
 */
import type { FriendEntryStatus } from './friendEntryStatus';

export interface FriendEntry {
  id: number;
  userId: number;
  friendId: number;
  friendUsername: string;
  friendCfHandle: string;
  friendRating: number;
  friendRank: string;
  status: FriendEntryStatus;
  isOnline?: boolean;
}