# lib/api-zod/src/generated/types/problem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CodeClash - Real-time competitive coding battle platform
 * OpenAPI spec version: 0.1.0
 */

export interface Problem {
  id: number;
  contestId: number;
  problemIndex: string;
  title: string;
  rating: number;
  tags: string[];
  cfUrl: string;
  /** @nullable */
  solvedCount?: number | null;
}